\name{chromOffsets}
\alias{chromOffsets}
\title{defines chromosome offsets for plotting multi chromosomal plot (eg plotManhattan)}
\usage{
chromOffsets(genome, space = 0.01)
}
\arguments{
  \item{genome}{A genome object to be used (2 columns:
  column 1 = chromosome name, column 2 = length of
  chromosome)}

  \item{space}{the space in between each chromosome as a
  fraction of the width of the plot}
}
\description{
defines chromosome offsets for plotting multi chromosomal
plot (eg plotManhattan)
}

