% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_counts.R
\name{collapse_counts}
\alias{collapse_counts}
\title{collapse_counts}
\usage{
collapse_counts(l2fc)
}
\description{
collapses l2fc values and computes MAD/sqrt(n) metrics for treatment conditions
}
\details{
@param l2fc - l2fc table with MAD/sqrt(n) metric for control condition
@returns - collapsed_counts
@export
}
