% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_l2fc.R
\name{compute_l2fc}
\alias{compute_l2fc}
\title{compute_l2fc}
\usage{
compute_l2fc(normalized_counts, control_type)
}
\arguments{
\item{normalized_counts}{\itemize{
\item table with normalized_n column and control_sample column that designates the
name of the control sample for each treatment sample
}}

\item{control_type}{\itemize{
\item string that denotes which compute samples to compute log fold change. Matches trt_type field
}}
}
\value{
l2fc data.frame with l2fc column
}
\description{
takes normalized counts and computes log-fold change values as compared to the annotated control columns
}
