% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_raw_reads.R
\name{filter_raw_reads}
\alias{filter_raw_reads}
\title{filter raw reads}
\usage{
filter_raw_reads(
  raw_counts,
  sample_meta,
  cell_line_meta,
  cell_set_meta,
  CB_meta,
  id_cols = c("cell_set", "treatment", "dose", "dose_unit", "day", "bio_rep",
    "tech_rep")
)
}
\arguments{
\item{raw_counts}{\itemize{
\item an unfiltered counts table
}}

\item{sample_meta}{\itemize{
\item the sample metadata for the particular experiment. Must follow the given set of guidelines for metadata
}}

\item{cell_line_meta}{\itemize{
\item master metadata of cell lines
}}

\item{cell_set_meta}{\itemize{
\item master metdata of cell sets and their contents
}}

\item{CB_meta}{\itemize{
\item master metdata of control barcodes, their sequences, and their doses
}}

\item{id_cols}{\itemize{
\item vector of column names used to generate unique profile_id for each sample.
cell_set,treatment,dose,dose_unit,day,bio_rep,tech_rep by default
}}
}
\value{
\itemize{
\item list with the following elements
#' \itemize{
\item filtered_table df of annotated readcounts
\item qc_table: QC table of index_purity and cell_line_purity
}
}
}
\description{
takes the raw readcount table and filters for expected indices and cell lines
using the given metadata. QC metrics are returned as a data.frame
}
