% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_biomarkers.R
\name{generate_biomarkers}
\alias{generate_biomarkers}
\title{generate_biomarkers}
\usage{
generate_biomarkers(collapsed_values)
}
\arguments{
\item{collapsed_values:}{collapsed_l2fc values}
}
\value{
Biomarker object containing lin_out, rf_out and disc_out tables
}
\description{
collapses filtered normalized counts and computes MAD/sqrt(n) metrics.
cell lines with MAD/sqrt(n) > 0.5 are filtered out, and file with filtered out cell lines is written.
log10(median counts) vs. MAD/sqrt(n) graph is saved, and collapsed filtered count table is returned
}
