% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlighthic.R
\name{highlighthic}
\alias{highlighthic}
\title{highlight loci on a HiC plot}
\usage{
highlighthic(x, y, labeltype = "bin", flip = FALSE, diameter = 0.025,
  lwd = 1, resolution = NULL, plottype = "triangle", half = "above",
  ...)
}
\arguments{
\item{x}{bin or midpoint of one interacting loci}

\item{y}{bin or midpoint of other interacting loci}

\item{labeltype}{the units for x and y options are 'bin' or 'bp'}

\item{flip}{TRUE/FALSE whether plot should be flipped over the x-axis}

\item{diameter}{of circle as a fraction of the x-axis}

\item{lwd}{The line width, a positive number, defaulting to 1.}

\item{resolution}{the width in bp of each pixel}

\item{plottype}{options are "triangle" or "square"}

\item{half}{if plottype is set to "square" should the highlight be "above" or "below" the diagonal}
}
\description{
highlight loci on a HiC plot
}
\examples{

}

