% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{normalize}
\usage{
normalize(X, barcodes)
}
\arguments{
\item{X}{\itemize{
\item dataframe of annotated readcounts that must include the following columns:
log_n or n: raw readcounts or log10(n) of read counts. Computes log_n if not present
log_dose: log10 of dose at which control barcode was spiked in, if applicable
sample_ID: some identifier that distinguishes between each sample
Name: contains the name of the control barcode that the read corresponds to, or NA
}}

\item{barcodes}{\itemize{
\item a vector of control barcode Name identifiers
}}
}
\value{
Table with counts normalized to control barcodes
}
\description{
takes a filtered dataframe of raw read counts and normalizes
counts using control barcodes
}
