% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_images.R
\name{QC_images}
\alias{QC_images}
\title{QC_images}
\usage{
QC_images(filtered_counts, cell_set_meta, out = NA)
}
\arguments{
\item{filtered_counts}{\itemize{
\item dataframe of annotated readcounts that must include the following columns:
n: raw readcounts
profile_id: string unique to each sample as defined by filter_counts method
Name: name of the control barcode that the read corresponds to, or NA (if read is cell line)
CCLE_name: name of the cell line that the read corresponds to, or NA (if read is control barcode)
cell_set: string identifier of cell set expected in a given sample, must match a cell set found in cell_set_meta
}}

\item{cell_set_meta}{\itemize{
\item a metadata dataframe that contains a mapping from cell set names (e.g. CS5) to
lists of LUAs in that cell set separated by semicolons
}}

\item{out}{\itemize{
\item the filepath to the folder in which QC images are meant to be saved, NA by default and
images are saved in the working directory
}}
}
\value{
\itemize{
\item NA, QC images are written out to the specified folder
}
}
\description{
takes a filtered dataframe of raw read counts and generates and writes out QC images to assess
overall quality of project data
}
