% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_DE.R
\name{run_DE}
\alias{run_DE}
\title{run_DE}
\usage{
run_DE(
  filtered_counts,
  sample_cols = c("cell_set", "treatment", "dose", "dose_unit", "day", "bio_rep"),
  sig_cols = c("cell_set", "treatment", "dose", "dose_unit", "day"),
  control_type = "negcon"
)
}
\arguments{
\item{filtered_counts}{\itemize{
\item dataframe of annotated readcounts that must include the following columns:
n: raw readcounts
trt_type: spcification of sample type (e.g. whether a sample is a control or not)
CCLE_name: contains the name of the cell line that the read corresponds to, or NA
Name: contains the name of the control barcode that the read corresponds to, or NA
all columns specified in sample_cols and sig_cols
}}

\item{sample_cols}{\itemize{
\item a vector of column names denoting which values specify each individual sample
(by default: cell_set, treatment, dose, dose_unity, day, and bio_rep)
}}

\item{sig_cols}{\itemize{
\item a vector of column names denoting which values specify each individual signature
(by default: cell_set, treatment, dose, dose_unity, and day)
}}

\item{control_type}{\itemize{
\item character value specifying which samples to use as controls. Must match a value in the
trt_type column
}}
}
\value{
Table with logfold change values, standard errors, and significance values for all signatures
}
\description{
takes a filtered dataframe of raw read counts and applies DEseq to generate
logfold change, standard errors, and significance values
}
