% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sortChrom.R
\name{sortChrom}
\alias{sortChrom}
\title{sort chromosome files by chom name}
\usage{
sortChrom(genome)
}
\arguments{
\item{genome}{A genome object to be used (2 columns: column 1 = chromosome name, column 2 = length of chromosome)}
}
\description{
sort chromosome files by chom name
}

