% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sushi.R
\docType{package}
\name{Sushi-package}
\alias{Sushi}
\alias{Sushi-package}
\title{Sushi: Tools for visualizing genomics data}
\description{
Flexible, quantitative, and integrative genomic visualizations for
    publication-quality multi-panel figures
}
