% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasetsinfo.R
\docType{data}
\name{Sushi_5C.bedpe}
\alias{Sushi_5C.bedpe}
\title{Sushi_5C.bedpe}
\format{bedpe format}
\source{
Sanyal, A., Lajoie, B. R., Jain, G. & Dekker, J. The long-range interaction landscape of gene promoters. Nature 489, 109-113 (2012).
}
\usage{
Sushi_5C.bedpe
}
\description{
This data set list the genomic locations of 5C interactions in multiple cell lines
with coordinates based on the NCBI36 / hg18 genome build.
}

