\docType{data}
\name{Sushi_DNaseI.bedgraph}
\alias{Sushi_DNaseI.bedgraph}
\title{Sushi_DNaseI.bedgraph}
\format{bedgraph format}
\source{
Neph, S. et al. An expansive human regulatory lexicon
encoded in transcription factor footprints. Nature 489,
83-90 (2012).
}
\usage{
Sushi_DNaseI.bedgraph
}
\description{
This data set describes read depths across the genomce
resulting from a DNaseI hypersensitivity experiment in K562
cells with coordinates based on the NCBI36 / hg18 genome
build.
}

