\docType{data}
\name{Sushi_GWAS.bed}
\alias{Sushi_GWAS.bed}
\title{Sushi_GWAS.bed}
\format{bed format}
\source{
Ehret, G. B. et al. Genetic variants in novel pathways
influence blood pressure and cardiovascular disease risk.
Nature 478, 103-109 (2011).
}
\usage{
Sushi_GWAS.bed
}
\description{
Bed data representing results from a GWAS study of blood
pressure and cardiovascular disease risk with coordinates
based on the NCBI36 / hg18 genome build.
}

