% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasetsinfo.R
\docType{data}
\name{Sushi_hg18_genome}
\alias{Sushi_hg18_genome}
\title{Sushi_hg18_genome}
\format{two columns (column 1 = chromosome name, column 2 = length of chromosome)}
\source{
http://www.biomart.org/ and Consortium, T. E. P. An integrated encyclopedia of DNA elements in the human genome. Nature 489, 57-74 (2012).
}
\usage{
Sushi_hg18_genome
}
\description{
This data set describes the length of human chromosomes according to the NCBI36 / hg18 genome build.
}

