\docType{data}
\name{Sushi_RNASeq_K562.bedgraph}
\alias{Sushi_RNASeq_K562.bedgraph}
\title{Sushi_RNASeq_K562.bedgraph}
\format{bedgraph format}
\source{
Consortium, T. E. P. An integrated encyclopedia of DNA
elements in the human genome. Nature 489, 57-74 (2012).

Consortium, T. E. P. An integrated encyclopedia of DNA
elements in the human genome. Nature 489, 57-74 (2012).
}
\usage{
Sushi_RNASeq_K562.bedgraph

Sushi_RNASeq_K562.bedgraph
}
\description{
Bedgraph data representing RNA-seq dat from K562 with
coordinates based on the NCBI36 / hg18 genome build.

Bedgraph data representing RNA-seq dat from K562 with
coordinates based on the NCBI36 / hg18 genome build.
}

