\docType{data}
\name{Sushi_transcripts.bed}
\alias{Sushi_transcripts.bed}
\title{Sushi_transcripts.bed}
\format{bed format}
\source{
http://www.biomart.org/ and Consortium, T. E. P. An
integrated encyclopedia of DNA elements in the human
genome. Nature 489, 57-74 (2012).
}
\usage{
Sushi_transcripts.bed
}
\description{
Bed data representing human transcripts and their
expression in K562 cells with coordinates based on the
NCBI36 / hg18 genome build.
}

