% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_counts.R
\name{collapse_counts}
\alias{collapse_counts}
\title{collapse_counts}
\usage{
collapse_counts(l2fc)
}
\description{
collapses filtered normalized counts and computes MAD/sqrt(n) metrics.
cell lines with MAD/sqrt(n) > 0.5 are filtered out, and file with filtered out cell lines is written.
log10(median counts) vs. MAD/sqrt(n) graph is saved, and collapsed filtered count table is returned
}
\details{
@param l2fc - normalized counts with bad replicates already filtered out
@returns collapsed_counts
}
