\docType{data}
\name{Sushi_ChIPExo_CTCF.bedgraph}
\alias{Sushi_ChIPExo_CTCF.bedgraph}
\title{Sushi_ChIPExo_CTCF.bedgraph}
\format{bedgraph format}
\source{
Rhee, H. S. & Pugh, B. F. Comprehensive genome-wide
protein-DNA interactions detected at single-nucleotide
resolution. Cell 147, 1408-1419 (2011).
}
\usage{
Sushi_ChIPExo_CTCF.bedgrap
}
\description{
This data set describes read depths across the genomce
resulting from a CTCF ChIP Exo experiment in K562 cells
with coordinates based on the NCBI36 / hg18 genome build.
}

